unsatisfying removal of typed classes. Need to update BaseEstimator so that can get rid of run and predict, with GridSearch still work

+++ src/main/scala/com/kindred/sclearn/metrics/ClassificationMetrics.scala
@@ -5,7 +5,7 @@ import breeze.linalg._
 
 object ClassificationMetrics {
 
-  def Accuracy(yPred: DenseVector[Int], y: DenseVector[Int]): Double = {
+  def Accuracy(yPred: DenseVector[Double], y: DenseVector[Double]): Double = {
     // TODO: check what is returned by :== operator
     (yPred :== y).length / yPred.length
   }
